updated for VS2015 as a static library.

+++ src/tuner_impl.cc
@@ -192,7 +192,7 @@ void TunerImpl::Tune() {
         if (tuning_result.time == std::numeric_limits<double>::max()) {
           tuning_result.time = 0.0;
           PrintResult(stdout, tuning_result, kMessageFailure);
-          tuning_result.time = std::numeric_limits<double>::max();
+          tuning_result.time = std::numeric_limits<float>::max();
         }
         else if (!tuning_result.status) {
           PrintResult(stdout, tuning_result, kMessageWarning);
